net: stmmac: Mac2mac-Update Fixed-link speed property in dtsi for partition case
In case of Mac2mac when speed property is passed from the partition
update fixed-link speed property in the dtsi instead of updating in
the driver.

+++ drivers/net/ethernet/stmicro/stmmac/dwmac-qcom-ethqos.h
@@ -511,7 +511,7 @@ struct ip_params {
 struct mac_params {
 	phy_interface_t eth_intf;
 	bool is_valid_eth_intf;
-	unsigned long link_speed;
+	unsigned int link_speed;
 };
 
 int ethqos_enable_serdes_consumers(struct qcom_ethqos *ethqos);
